Moved SecondaryTypeEdge to upper class
