+++ cpg-core/src/main/kotlin/de/fraunhofer/aisec/cpg/graph/declarations/TypeParamDeclaration.kt
@@ -27,7 +27,7 @@ package de.fraunhofer.aisec.cpg.graph.declarations
 
 import de.fraunhofer.aisec.cpg.graph.AST
 import de.fraunhofer.aisec.cpg.graph.HasDefault
-import de.fraunhofer.aisec.cpg.graph.HasLegacyType.SecondaryTypeEdge
+import de.fraunhofer.aisec.cpg.graph.types.SecondaryTypeEdge
 import de.fraunhofer.aisec.cpg.graph.types.Type
 import java.util.*
 import org.neo4j.ogm.annotation.Relationship
@@ -37,8 +37,8 @@ class TypeParamDeclaration : ValueDeclaration(), SecondaryTypeEdge, HasDefault<T
     /**
      * TemplateParameters can define a default for the type parameter Since the primary type edge
      * points to the ParameterizedType, the default edge is a secondary type edge. Therefore, the
-     * TypeResolver requires to implement the [HasType.SecondaryTypeEdge] to be aware of the edge to
-     * be able to merge the type nodes.
+     * TypeResolver requires to implement the [SecondaryTypeEdge] to be aware of the edge to be able
+     * to merge the type nodes.
      */
     @Relationship(value = "DEFAULT", direction = Relationship.Direction.OUTGOING)
     @AST
